Hack The UmmIt - Info/pyproject.toml
[project]
name = "pycon-ethical-hacking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "python-dotenv>=1.2.1",
]
